Corrections et complements INSTALLATION AUTOMATIQUE #50

+++ install_db.sh
@@ -56,9 +56,9 @@ then
 
     sudo -n -u postgres -s psql -d $db_name -c "CREATE SCHEMA synthese AUTHORIZATION $admin_pg;
                                                 CREATE SCHEMA taxonomie AUTHORIZATION $admin_pg;
-                                                CREATE SCHEMA utilisateurs AUTHORIZATION $admin_pg;
-                                                CREATE SCHEMA meta AUTHORIZATION $admin_pg;
                                                 CREATE SCHEMA atlas AUTHORIZATION $admin_pg;"
+
+
     
     #ajout du shape des limite du territoire
     ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 data/ref/emprise_territoire_3857.shp $limit_shp
@@ -67,32 +67,41 @@ then
     #Creation de l'index GIST sur la layer territoire
     export PGPASSWORD=$admin_pg_pass; psql -h $db_host -U $admin_pg -d $db_name -c "CREATE INDEX index_gist_t_layer_territoire
                                                                                     ON atlas.t_layer_territoire
-                                                                                    USING gist(geom);"
+                                                                                    USING gist(geom);
+                                                                                    ALTER TABLE atlas.t_layer_territoire RENAME COLUMN geom TO the_geom;"
+        #conversion en json
+    rm  -f ../../static/custom/territoire.json
+    ogr2ogr -f "GeoJSON" -t_srs "EPSG:4326" ../../static/custom/territoire.json $limit_shp
 
 
-    ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 ./data/ref/communes_3857.shp $communes_shp
-    export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I ./data/ref/communes_3857.shp atlas.l_communes | psql -h $db_host -U $admin_pg $db_name
-    psql -h $db_host -U $admin_pg $db_name -c "ALTER TABLE atlas.l_communes RENAME COLUMN $colonne_nom_commune TO commune_maj;
-                                               ALTER TABLE atlas.l_communes RENAME COLUMN $colonne_insee TO insee;
-                                               CREATE INDEX index_gist_t_layers_communes
-                                                ON atlas.l_communes USING gist (geom); "
-    rm ./data/ref/communes_3857.*
+    if [ import_commune_shp ]; then
+        ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 ./data/ref/communes_3857.shp $communes_shp
+        export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I ./data/ref/communes_3857.shp atlas.l_communes | psql -h $db_host -U $admin_pg $db_name
+        psql -h $db_host -U $admin_pg $db_name -c "ALTER TABLE atlas.l_communes RENAME COLUMN $colonne_nom_commune TO commune_maj;
+                                                   ALTER TABLE atlas.l_communes RENAME COLUMN $colonne_insee TO insee;
+                                                   ALTER TABLE atlas.l_communes RENAME COLUMN geom TO the_geom;
+                                                   CREATE INDEX index_gist_t_layers_communes
+                                                    ON atlas.l_communes USING gist (the_geom); "
+        rm ./data/ref/communes_3857.*
+    fi
 
 
 
     echo "Création de la structure de la base..."
 
-    #Creation des foreign data
-    sudo cp data/atlas_geonature.sql /tmp/atlas_geonature.sql
-    sudo sed -i 's/TO myuser;/TO geonatadmin;/g' /tmp/atlas_geonature.sql
-    export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -f /tmp/atlas_geonature.sql  &>> log/install_db.log
+    if [ $geonature_user ]; then
+        #Creation des foreign data
+        sudo cp data/atlas_geonature.sql /tmp/atlas_geonature.sql
+        sudo sed -i 's/TO myuser;/TO geonatadmin;/g' /tmp/atlas_geonature.sql
+        export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -f /tmp/atlas_geonature.sql  &>> log/install_db.log
+    fi
 
     # Creation des VM
     sudo cp data/atlas.sql /tmp/atlas.sql
     sudo sed -i "s/WHERE id_attribut IN (100, 101, 102, 103);$/WHERE id_attribut  IN ($attr_desc, $attr_commentaire, $attr_milieu, $attr_chorologie);/" /tmp/atlas.sql
     sudo sed -i "s/FROM layers.l_communes$/FROM $table_commune/" /tmp/atlas.sql
-    sudo sed -i "s/current_date -15;$/current_date -$time/" /tmp/atlas.sql
-    sudo sed -i "s/current_date +15;$/current_date +$time/" /tmp/atlas.sql
+    sudo sed -i "s/date - 15$/date - $time/" /tmp/atlas.sql
+    sudo sed -i "s/date + 15$/date - $time/" /tmp/atlas.sql
 
     export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -f /tmp/atlas.sql  &>> log/install_db.log
 
@@ -114,38 +123,35 @@ then
 
     rm -f L93*.dbf L93*.prj L93*.sbn L93*.sbx L93*.shp L93*.shx
 
-    unzip L93_1K.zip 
-    unzip L93_5K.zip 
-    unzip L93_10K.zip
+
 
     if [ $metropole ]; then
-    #ajout des mailles non decoupees
-    ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_1.shp L93_1x1.shp
-    ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_5.shp L93_5K.shp
-    ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_10.shp L93_10K.shp
+        #unzip mailles
+        unzip L93_1K.zip 
+        unzip L93_5K.zip 
+        unzip L93_10K.zip
+        #ajout des mailles non decoupees
+        ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_1.shp L93_1x1.shp
+        ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_5.shp L93_5K.shp
+        ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 mailles_10.shp L93_10K.shp
+
+        export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I mailles_$taillemaille.shp atlas.t_mailles_$taillemaille | psql -h $db_host -U $admin_pg $db_name
+    
 
 
-        export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I mailles_1.shp atlas.t_mailles_1 | psql -h $db_host -U $admin_pg $db_name
-        export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I mailles_5.shp atlas.t_mailles_5 | psql -h $db_host -U $admin_pg $db_name
-        export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I mailles_10.shp atlas.t_mailles_10 | psql -h $db_host -U $admin_pg $db_name
+        rm mailles_1.* mailles_5.* mailles_10.* L93_1x1.* L93_5K.* L93_10K.*
 
-        rm mailles_1.* mailles_5.* mailles_10.*
-
-
-        #conversion en json
-        rm  -f ../../static/custom/territoire.json
-        ogr2ogr -f "GeoJSON" -t_srs "EPSG:4326" ../../static/custom/territoire.json $limit_shp
 
         cd ../../
         echo $taillemaille
         # Creation de la table t_maille_territoire avec la taille de maille passée en parametre
         export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -c "CREATE TABLE atlas.t_mailles_territoire as
-                                                    SELECT m.geom, ST_AsGeoJSON(st_transform(t.geom, 4326))
+                                                    SELECT m.geom AS the_geom, ST_AsGeoJSON(st_transform(t.the_geom, 4326))
                                                     FROM atlas.t_mailles_"$taillemaille" m, atlas.t_layer_territoire t
-                                                    WHERE ST_Intersects(m.geom, t.geom);
+                                                    WHERE ST_Intersects(m.geom, t.the_geom);
                                                     CREATE INDEX index_gist_t_mailles_territoire
                                                     ON atlas.t_mailles_territoire
-                                                    USING gist (geom); 
+                                                    USING gist (the_geom); 
                                                     ALTER TABLE atlas.t_mailles_territoire
                                                     ADD COLUMN id_maille serial;
                                                     ALTER TABLE atlas.t_mailles_territoire
@@ -154,23 +160,25 @@ then
         ogr2ogr -f "ESRI Shapefile" -t_srs EPSG:3857 custom_mailles_3857.shp $chemin_custom_maille 
         export PGPASSWORD=$admin_pg_pass;shp2pgsql -W "LATIN1" -s 3857 -D -I custom_mailles_3857.shp atlas.t_mailles_custom | psql -h $db_host -U $admin_pg $db_name
 
-                export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -c "CREATE TABLE atlas.t_mailles_territoire as
-                                                    SELECT m.geom, ST_AsGeoJSON(st_transform(t.geom, 4326))
-                                                    FROM atlas.t_mailles_custom m, atlas.t_layer_territoire t
-                                                    WHERE ST_Intersects(m.geom, t.geom);
-                                                    CREATE INDEX index_gist_t_mailles_custom
-                                                    ON atlas.t_mailles_territoire
-                                                    USING gist (geom); 
-                                                    ALTER TABLE atlas.t_mailles_territoire
-                                                    ADD COLUMN id_maille serial;
-                                                    ALTER TABLE atlas.t_mailles_territoire
-                                                    ADD PRIMARY KEY (id_maille);"
+        export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -c "CREATE TABLE atlas.t_mailles_territoire as
+                                            SELECT m.geom AS the_geom, ST_AsGeoJSON(st_transform(t.the_geom, 4326))
+                                            FROM atlas.t_mailles_custom m, atlas.t_layer_territoire t
+                                            WHERE ST_Intersects(m.geom, t.the_geom);
+                                            CREATE INDEX index_gist_t_mailles_custom
+                                            ON atlas.t_mailles_territoire
+                                            USING gist (the_geom); 
+                                            ALTER TABLE atlas.t_mailles_territoire
+                                            ADD COLUMN id_maille serial;
+                                            ALTER TABLE atlas.t_mailles_territoire
+                                            ADD PRIMARY KEY (id_maille);"
+
+
     fi
 
 
     echo "Creation de la table des mailles..."
     #ajout de la tabme vm_mailles_observations
-    export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -f data/mailles.sql  &> log/install_mailles.log
+    export PGPASSWORD=$admin_pg_pass;psql -h $db_host -U $admin_pg -d $db_name -f data/observations_mailles.sql  &> log/install_mailles.log
 
     echo "Grant..."
     sudo cp data/grant.sql /tmp/grant.sql
